Create ayush_21bcon295_2.py
Tic-Tac-Toe Game :-
I used basic string & input methods along with scopes, while loop & for loop to create this program.
It works on 3x3 grid & coordinate system i.e, we have to input coordinates to make our move. The input should be in the form: row_number column_number i.e. 1 1 or 2 2, etc. Don't forget to give a space between them.
Some of its key features are that it checks userinput for wrong input i.e, It will raise warning if we give 'two two' as an input or an input which is out of range eg., 1 4.