Create ayush_21bcon295_4.py
Hangman
It is a classic word guessing game. 
You have 8 attempts to guess the word or you will be hanged!
Some of its features: It prompts users for any wrong input or for already guessed words.
It has a menu that allows users to either exit or play.
It counts for 8 wrong attempts after which the game is over and you are hanged.