Create ayush_21bcon295_6.py
Dominoes Game
It is a classic dominoes game.
It requires lots of brainstorming and is worth it. I learned a lot in depth about loops and functions through this.
The game is played between the computer and the player. Yes, I also coded an AI algorithm so that the computer makes automatic and intelligent moves. This is the main feature of this project - the AI.
Some important features:-
It is based on AI which means that the game is between the computer and a human. The computer makes automatic and intelligent moves.
It prompts the human player for any invalid input and keeps asking for input till a valid one is received.
It checks winning and draw conditions after every move and stops the game if conditions are met, otherwise the game continues.