A file transfer for windows one way sharing to any device
No need for pen drives when your work can be done by setting up a local webserver. Used python HTTP server to create a one-way gateway for sharing files with friends. Connect to the pc same hotspot created by the bat file and login in from the other device and enjoy the files shared. Can be used to provide the same files to many people at the same time i.e in tech-fest with a small upgrade in hardware. Use the stopHotspot.bat to close the hosted network.